feat(api): eliminate unknown casts in apiClient — 18 methods now use DTO types
- getEntity/getDocument/getEvidence: parseWithSchema validates API responses
- getEntityGraph: typed as GraphResponseDto via graph schema
- search: returns UnifiedSearchResponseDto instead of manual Record mapping
- getAllEntities, getInvestigations, getInvestigation, createInvestigation: schema-typed returns
- Annotations: use PublicDocumentAnnotation shared DTO
- Entity documents/related docs: typed as DocumentListItemDto[]
- Replaced 20+ unknown/Record casts with proper DTO types
- knip.json: added 10 known-used deps to ignoreDependencies
- scripts/seed_minimal.ts: fixed legacy entities.type reference
- Fixed 5 caller files for stricter type compatibility
- Production build passes all CI gates (type-check, knip, schema, budgets)

+++ src/server/db/documentsRepository.ts
@@ -692,7 +692,7 @@ export const documentsRepository = {
         SELECT entity_id, mention_context
         FROM entity_mentions
         WHERE document_id = $1
-          AND entity_id = ANY($2::int[])
+          AND entity_id = ANY($2::bigint[])
           AND mention_context IS NOT NULL
           AND mention_context != ''
         ORDER BY entity_id, id
